Move test to unit test
